bug fixes; fix flaky build
